Revise and enhance ELECTENG-332 notes structure
Reorganized and renamed main document for clarity. Introduced new introductory chapters detailing the organization and usage of the notes. Enhanced readability with new color settings and added `tcolorbox` package. Included complex frequency explanations and updated figures paths for better presentation. Aimed to improve the educational utility and navigability of the material for students.